refactor: create service.go
